backend/pyproject.toml
[tool.poetry]
name = "talk-space"
version = "0.1.0"
description = "Let's share our opinions"
authors = ["Liad Hossain <liadhossain@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
uvicorn = "^0.34.2"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
djangorestframework-simplejwt = "^5.5.0"
django-ratelimit = "^4.1.0"
django-blacklist = "^0.7.0"
pre-commit = "^4.2.0"
django-cors-headers = "^4.7.0"
pusher = "^3.3.3"
drf-spectacular = "^0.28.0"
django-redis = "^5.4.0"
taskiq-redis = "^1.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
